pyproject.toml
[project]
name = "myproject"
version = "0.1.0"
description = "A Python project"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.1",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "streamlit>=1.56.0",
    "tavily-python>=0.7.23",
    "uvicorn>=0.46.0",
]


[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[dependency-groups]
dev = [
    "mypy>=1.20.2",
    "pytest>=9.0.3",
    "ruff>=0.15.11",
    "ipykernel==6.30.1",
]
